Add files via upload
Qn 18
You have two friends who are speaking strange language. The language is the same expect that
you take the first of each word and put it on the end, then you add 'ay' to the end of that
word for example 'hello', = 'ellohay'
Take the sentence in english and turn it into the same sentence as the strange languge
Input: hello everyone this is TDC
Output: ohellay eeveryonay sthiay siay CTDay.